update to website
This website update helps transition the online Cydog Intensifies feature to allow for increased adoption and more mainstream use. Some terms have been changed. For instance, the license for the Cydog Intensifies code is now CC BY-NC-SA 4.0. This was done to increase innovation in the search engine space through open-source, non-commericial means. Oftentimes, code requiring innovation is held tightly by those seeking to over-monetize and code not requiring innovation is made free and publicly available for the world to use. It is about time we change this process. Not all code should be free, as developers need to make a living. However, there is virtually no reason on the planet to make search engine codes or algorithms proprietary, as good search features are based on data only those companies have and no one could hope to push them out of. Thank you for using Cydog Browser!